feat(core): moved firebase file to its expected location
