+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { AuthProvider } from "./features/auth/contexts/auth-provider";
 import { Router } from "./features/core/router";
-import { auth } from "./features/auth/firebase";
+import { auth } from "./features/core/firebase";
 import { BrowserRouter } from "react-router";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import { theme } from "./features/core/theme";
